fixed resize group object (#10456).
svn path=/trunk/koffice/; revision=123969

+++ kpresenter/kpgroupobject.h
@@ -87,7 +87,7 @@ public:
     virtual void doSpecificEffects( bool _specEffects, bool _onlyCurrStep = true );
 
 protected:
-    void updateSizes( float fx, float fy );
+    void updateSizes( double fx, double fy );
     void updateCoords( int dx, int dy );
 
     QPtrList<KPObject> objects;
